+++ spring.hibernate/src/main/java/com/example/dao/BankDao.java
@@ -3,7 +3,7 @@ package com.example.dao;
 import com.example.entity.Bank;
 
 public interface BankDao {
-	Bank findById(int id);
+	Bank findById(long id);
 	void save(Bank user);
 
 }
